Move Tileset tests to a separate file
